# openapi3_0.yaml
openapi: 3.0.0
info:
  title: Lost and Found System API
  version: 1.0.0
  description: API for managing lost and found items, auctions, and user accounts.
servers:
  - url: https://api.lostandfound.com/v1
...
paths:
...
  /items/compare/{lostItemId}/{foundItemId}:
    get:
      summary: Compare a lost item with a found item
      tags:
        - Lost Items
        - Found Items
      parameters:
        - in: path
          name: lostItemId
          required: true
          schema:
            type: integer 
        - in: path
          name: foundItemId
          required: true
          schema:
            type: integer 
      responses:
        '200':
          description: Comparison result between lost and found item
          content:
            application/json:
              schema:
                type: object
                properties:
                  similarities:
                    type: array
                    items:
                      type: string
                  differences:
                    type: array
                    items:
                      type: string
  /items/found:
    get:
      tags:
        - Found Items
      summary: Search for found items
      parameters:
        - in: query
          name: description
          schema:
            type: string
          required: false
          description: Description of the item
      responses:
        '200':
          description: List of found items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FoundItem'
...
  /police/posts:
    post:
      tags:
        - Police Operations
      summary: Register a new police post
      description: This endpoint registers a new police post by providing the required address (morada) information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - morada
              properties:
                morada:
                  type: string
                  description: The address of the new police post.
      responses:
        '201':
          description: Police post created successfully. Returns the ID of the newly created post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        '400':
          description: Invalid input. The "morada" field is missing or is not a valid string.
        '500':
          description: Server error while creating the police post.
  /police/members:
    post:
      tags:
        - Police Operations
      summary: Register a new police member
      description: Adds a new member to the police database. The system automatically generates a unique ID for each new member.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - posto_policia
              properties:
                nome:
                  type: string
                  description: The full name of the police member.
                posto_policia:
                  type: integer
                  description: The reference ID to an existing police station record.
                historico_policia:
                  type: string
                  description: A JSON formatted string detailing the police member's history.
      responses:
        '201':
          description: Police member registered successfully. Returns the ID of the newly registered member.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        '400':
          description: Invalid input, required fields are missing or incorrect type of data was provided.
        '500':
          description: Internal server error. Indicates a server-side error that prevented the member from being registered.
  /police/items/found/register:
    post:
      tags:
        - Police Operations
      summary: Register a found item by police
      description: This endpoint allows police to register items that have been found. It requires detailed item information and returns an ID of the newly registered item if successful.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoundItem'
      responses:
        '201':
          description: Found item registered successfully. Returns the ID of the newly registered item.
        '400':
          description: Invalid input. One or more of the required fields are missing or malformed.
        '500':
          description: Internal server error. Indicates a server-side error that prevented the item from being registered.
  /police/items/{itemId}/claim:
    put:
      tags:
        - Police Operations
      summary: Claim a found item
      description: Claims an active item as found by updating its status.
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
          description: The unique identifier of the item to claim.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                claimantId:
                  type: integer
                  description: The unique identifier of the user claiming the item.
      responses:
        '200':
          description: Item claimed successfully.
        '400':
          description: Invalid input, itemId or claimantId must be integers.
        '404':
          description: Item not found or already claimed.
        '500':
          description: Server error while claiming the item.
...
components:
  schemas:
...
    FoundItem:
      type: object
      properties:
        itemId: 
          type: integer 
        description:
          type: string
        location:
          type: string
        dateFound:
          type: string
          format: date
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
security:
  - ApiKeyAuth: []